docker_compose/django/pyproject.toml
[tool.poetry]
name = "admin-panel"
version = "0.1.0"
description = ""
authors = ["Dmitry Murzin <dima@murzin.life>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
django = "4.2.5"
django-split-settings = "1.2.0"
psycopg2 = "2.9.7"
types-psycopg2 = "2.9.21.13"
uwsgi = "2.0.22"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "4.2.0"
pytest = "7.4.2"
python-dateutil = "2.8.2"
black = "23.9.1"
mypy = "1.5.1"
ruff = "0.0.291"
flake8 = "6.1.0"
flake8-blind-except = "0.2.1"
flake8-broken-line = "1.0.0"
flake8-bugbear = "23.9.16"
flake8-builtins = "2.1.0"
flake8-class-attributes-order = "0.1.3"
flake8-cognitive-complexity = "0.1.0"
flake8-commas = "2.1.0"
flake8-comprehensions = "3.14.0"
flake8-debugger = "4.1.2"
flake8-eradicate = "1.5.0"
flake8-functions = "0.0.8"
flake8-isort = "6.1.0"
flake8-mutable = "1.2.0"
flake8-print = "5.0.0"
flake8-pytest = "1.4"
flake8-pytest-style = "1.7.2"
flake8-quotes = "3.3.2"
flake8-string-format = "0.3.0"
flake8-variables-names = "0.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E501"]
